Add custom VCP code support for brightness control
Introduces the ability for users to specify a custom VCP code for brightness control per monitor. Updates the UI to allow input of a custom VCP code, updates monitor state and logic to use the custom code if set, and adds relevant localization strings. Also ensures known VCP codes are not mutated when merging with user settings.